% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session_details.R
\name{session_details}
\alias{session_details}
\title{Track session details}
\usage{
session_details()
}
\value{
A \code{session_info} object written to a \code{TXT} file ("~/session_info.txt") and a \code{session_diff} object saved to an \code{RDS} file ("~/session_diff.rds").
}
\description{
Print and compare details of the current and prior sessions using \code{\link[sessioninfo:session_info]{sessioninfo::session_info()}} and \code{\link[sessioninfo:session_diff]{sessioninfo::session_diff()}}.
}
